third_party/blink/web_tests/external/wpt/service-workers/service-worker/fetch-event-redirect.https-expected.txt

This is a testharness.js-based test.
Found 5 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Non-navigation, follow redirect, cors mode Request redirected to same-origin with credentials should succeed interception and response should be redirected
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
[FAIL] Non-navigation, follow redirect, same-origin mode Request redirected to same-origin with credentials should succeed interception and response should be redirected
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
[FAIL] Non-navigation, follow redirect, no-cors mode Request redirected to same-origin with credentials should succeed interception and response should be redirected
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
[FAIL] Non-navigation, follow redirect, no-cors mode Request redirected to no-cors with credentials should succeed interception and response should not be redirected
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
[FAIL] Non-navigation, follow redirect, no-cors mode Request redirected to cors with credentials should succeed interception and response should not be redirected
  promise_test: Unhandled rejection with value: object "TypeError: Failed to fetch"
Harness: the test ran to completion.
